Extend obscore-config migration with creation of obscore table.
This is a non-trivial part for which we have to use Registry and obscore
manager, with potential compatibility issues. Because schema depends on
the user-provided configuration, we need obscore manager for deducing
actual schema. And because of the foreign key to dataset table we need
dataset manager, which basically means we have to instantiate complete
registry.