Ignore documentation string mismatch in dimension validation
Users are having trouble running universe v5 -> v6 migrations because of the error generated for mismatched dimension universes.  There is a known mismatch of documentation strings in some older repositories that is spuriously throwing an error.

Add a special case to ignore this known mismatch so that users don't have to add special options to complete the migration.